download snapshot if it doesn't exist

+++ modules/snapshot.sh
@@ -13,12 +13,16 @@ snapshot_download()
 {
     local target=${snapshot_dir}/current
 
-    if [[ $(expr `date +%s` - `stat -c %Y $target`) -gt 900 ]]; then
-        if [[ -e $target ]]; then
-            rm -f ${target} >> $noah_log 2>&1
-        fi
-
+    if [ ! -f $target ]; then
         wget -nv ${snapshot} -O ${target} >> $noah_log 2>&1
+    else
+        if [[ $(expr `date +%s` - `stat -c %Y $target`) -gt 900 ]]; then
+            if [[ -e $target ]]; then
+                rm -f ${target} >> $noah_log 2>&1
+            fi
+
+            wget -nv ${snapshot} -O ${target} >> $noah_log 2>&1
+        fi
     fi
 }
